requirements.txt
numpy
pandas
scikit-learn
scipy
pyTsetlinMachine
matplotlib
zemberek-python     
zeyrek             
scikit-fuzzy
networkx